LayoutTests/platform/mac/media/encrypted-media/fps-encrypted-event-expected.txt


RUN(video.srcObject = mediaSource)
EVENT(sourceopen)
RUN(sourceBuffer = mediaSource.addSourceBuffer("video/mp4"))
RUN(sourceBuffer.appendBuffer(mediaData))
EVENT(encrypted)
EXPECTED (event.initData != 'null') OK
EXPECTED (event.initDataType == 'sinf') OK
END OF TEST
